app/requirements.txt
flask
torch 
numpy
flask_cors
pillow
roboflow
pandas
dill
gunicorn
ultralytics